+++ src/wasm/body.rs
@@ -2,7 +2,7 @@
 use super::multipart::Form;
 /// dox
 use bytes::Bytes;
-use js_sys::Uint8Array;
+use js_sys::{Array, Uint8Array};
 use std::fmt;
 use wasm_bindgen::JsValue;
 
@@ -17,7 +17,6 @@ pub struct Body {
     inner: Inner,
 }
 
-#[derive(Clone)]
 enum Inner {
     Bytes(Bytes),
     #[cfg(feature = "multipart")]
@@ -41,7 +40,9 @@ impl Body {
         match &self.inner {
             Inner::Bytes(body_bytes) => {
                 let body_bytes: &[u8] = body_bytes.as_ref();
-                let body_array: Uint8Array = body_bytes.into();
+                let body_uint8_array: Uint8Array = body_bytes.into();
+                let body_array = Array::new();
+                body_array.push(&body_uint8_array);
                 let js_value: &JsValue = body_array.as_ref();
                 Ok(js_value.to_owned())
             }
@@ -70,9 +71,13 @@ impl Body {
         }
     }
 
-    pub(crate) fn clone(&self) -> Body {
-        Self {
-            inner: self.inner.clone(),
+    pub(crate) fn try_clone(&self) -> Option<Body> {
+        match &self.inner {
+            Inner::Bytes(bytes) => Some(Self {
+                inner: Inner::Bytes(bytes.clone()),
+            }),
+            #[cfg(feature = "multipart")]
+            Inner::Multipart(_) => None,
         }
     }
 }